Update 1.3 Bug FIX

+++ Controllers/ProductsController.cs
@@ -1,24 +1,29 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System;
+using System.IO;
+using System.Linq;
 using UsersApp.Data;
 using UsersApp.Models;
+using UsersApp.ViewModels;
 
 namespace UsersApp.Controllers
 {
     [Authorize(Roles = "Admin")]
     public class ProductsController : Controller
     {
-        private readonly AppDbContext context;
-        private readonly IWebHostEnvironment environment;
+        private readonly AppDbContext _context;
+        private readonly IWebHostEnvironment _environment;
 
         public ProductsController(AppDbContext context, IWebHostEnvironment environment)
         {
-            this.context = context;
-            this.environment = environment;
+            _context = context;
+            _environment = environment;
         }
+
         public IActionResult Index()
         {
-            var products = context.Products.OrderByDescending(p => p.Id).ToList();
+            var products = _context.Products.OrderByDescending(p => p.Id).ToList();
             return View(products);
         }
 
@@ -26,6 +31,7 @@ namespace UsersApp.Controllers
         {
             return View();
         }
+
         [HttpPost]
         public IActionResult Create(ProductDto productDto)
         {
@@ -35,13 +41,15 @@ namespace UsersApp.Controllers
             }
             if (!ModelState.IsValid)
             {
-                return View(productDto); 
-                
+                return View(productDto);
             }
 
-            string newFileName = DateTime.Now.ToString("yyyyMMddHHmmssfff");
-            newFileName += Path.GetExtension(productDto.ImageFile!.FileName);
-            string imageFullPath = environment.WebRootPath + "/products" + newFileName;
+            string newFileName = DateTime.Now.ToString("yyyyMMddHHmmssfff")
+                                 + Path.GetExtension(productDto.ImageFile.FileName);
+            string folderPath = Path.Combine(_environment.WebRootPath, "products");
+            Directory.CreateDirectory(folderPath);
+            string imageFullPath = Path.Combine(folderPath, newFileName);
+
             using (var stream = System.IO.File.Create(imageFullPath))
             {
                 productDto.ImageFile.CopyTo(stream);
@@ -58,20 +66,20 @@ namespace UsersApp.Controllers
                 CreatedAt = DateTime.Now
             };
 
-            context.Products.Add(product);
-            context.SaveChanges();
+            _context.Products.Add(product);
+            _context.SaveChanges();
 
-                return RedirectToAction("Index", "Products");
+            return RedirectToAction("Index", "Products");
         }
 
         public IActionResult Edit(int id)
         {
-            var product = context.Products.Find(id);
-            if(product == null)
+            var product = _context.Products.Find(id);
+            if (product == null)
             {
                 return RedirectToAction("Index", "Products");
             }
-            var productDto = new ProductDto()
+            var productDto = new ProductDto
             {
                 Name = product.Name,
                 Brand = product.Brand,
@@ -90,7 +98,7 @@ namespace UsersApp.Controllers
         [HttpPost]
         public IActionResult Edit(int id, ProductDto productDto)
         {
-            var product = context.Products.Find(id);
+            var product = _context.Products.Find(id);
             if (product == null)
             {
                 return RedirectToAction("Index", "Products");
@@ -107,46 +115,53 @@ namespace UsersApp.Controllers
             string newFileName = product.ImageFileName;
             if (productDto.ImageFile != null)
             {
-                newFileName = DateTime.Now.ToString("yyyyMMddHHmmssff");
-                newFileName += Path.GetExtension(productDto.ImageFile.FileName);
+                newFileName = DateTime.Now.ToString("yyyyMMddHHmmssff")
+                              + Path.GetExtension(productDto.ImageFile.FileName);
+                string folderPath = Path.Combine(_environment.WebRootPath, "products");
+                Directory.CreateDirectory(folderPath);
+                string imageFullPath = Path.Combine(folderPath, newFileName);
 
-                string imageFullPath = environment.WebRootPath + "/products/" + newFileName;
                 using (var stream = System.IO.File.Create(imageFullPath))
                 {
                     productDto.ImageFile.CopyTo(stream);
                 }
 
-                string oldImageFullPath = environment.WebRootPath + "/products/" + product.ImageFileName;
-                System.IO.File.Delete(oldImageFullPath);
+            
+                string oldImageFullPath = Path.Combine(_environment.WebRootPath, "products", product.ImageFileName);
+                if (System.IO.File.Exists(oldImageFullPath))
+                {
+                    System.IO.File.Delete(oldImageFullPath);
+                }
             }
 
             product.Name = productDto.Name;
             product.Brand = productDto.Brand;
             product.Category = productDto.Category;
             product.Price = productDto.Price;
-            product.Description = product.Description;
+            product.Description = productDto.Description;
             product.ImageFileName = newFileName;
 
-            context.SaveChanges();
+            _context.SaveChanges();
 
             return RedirectToAction("Index", "Products");
         }
 
         public IActionResult Delete(int id)
         {
-            var product = context.Products.Find(id);
+            var product = _context.Products.Find(id);
             if (product == null)
             {
                 return RedirectToAction("Index", "Products");
             }
-            string imageFullPath = environment.WebRootPath + "/products/" + product.ImageFileName;
-            System.IO.File.Delete(imageFullPath);
-            context.Products.Remove(product);
-            context.SaveChanges(true);
+            string imageFullPath = Path.Combine(_environment.WebRootPath, "products", product.ImageFileName);
+            if (System.IO.File.Exists(imageFullPath))
+            {
+                System.IO.File.Delete(imageFullPath);
+            }
+            _context.Products.Remove(product);
+            _context.SaveChanges();
 
             return RedirectToAction("Index", "Products");
         }
-
-       
     }
 }
